+++ dixitabot/db/__init__.py
@@ -1,0 +1,2 @@
+from .chatsdb import *
+from .usersdb import *
