+++ dixitabot/modules/helpers/__init__.py
@@ -1,0 +1,22 @@
+from typing import Callable
+
+from pyrogram.enums import ChatMemberStatus
+from pyrogram.types import Message
+
+from dixitabot import OWNER, app, NexiChat
+
+
+def is_admins(func: Callable) -> Callable:
+    async def non_admin(c: NexiChat, m: Message):
+        if m.from_user.id == OWNER:
+            return await func(c, m)
+
+        admin = await c.get_chat_member(m.chat.id, m.from_user.id)
+        if admin.status in [ChatMemberStatus.OWNER, ChatMemberStatus.ADMINISTRATOR]:
+            return await func(c, m)
+
+    return non_admin
+
+
+from .inline import *
+from .read import *
