+++ dixitabot/modules/helpers/assets.py
@@ -1,0 +1,40 @@
+# Centralized Assets Handlers
+
+# Random Start Images
+IMG = [
+    "https://graph.org/file/210751796ff48991b86a3.jpg",
+    "https://graph.org/file/7b4924be4179f70abcf33.jpg",
+    "https://graph.org/file/f6d8e64246bddc26b4f66.jpg",
+    "https://graph.org/file/63d3ec1ca2c965d6ef210.jpg",
+    "https://graph.org/file/9f12dc2a668d40875deb5.jpg",
+    "https://graph.org/file/0f89cd8d55fd9bb5130e1.jpg",
+    "https://graph.org/file/e5eb7673737ada9679b47.jpg",
+    "https://graph.org/file/2e4dfe1fa5185c7ff1bfd.jpg",
+    "https://graph.org/file/36af423228372b8899f20.jpg",
+    "https://graph.org/file/c698fa9b221772c2a4f3a.jpg",
+    "https://graph.org/file/61b08f41855afd9bed0ab.jpg",
+    "https://graph.org/file/744b1a83aac76cb3779eb.jpg",
+    "https://graph.org/file/814cd9a25dd78480d0ce1.jpg",
+    "https://graph.org/file/e8b472bcfa6680f6c6a5d.jpg",
+]
+
+# Random Stickers
+STICKER = [
+    "CAACAgUAAx0CYlaJawABBy4vZaieO6T-Ayg3mD-JP-f0yxJngIkAAv0JAALVS_FWQY7kbQSaI-geBA",
+    "CAACAgUAAx0CYlaJawABBy4rZaid77Tf70SV_CfjmbMgdJyVD8sAApwLAALGXCFXmCx8ZC5nlfQeBA",
+    "CAACAgUAAx0CYlaJawABBy4jZaidvIXNPYnpAjNnKgzaHmh3cvoAAiwIAAIda2lVNdNI2QABHuVVHgQ",
+]
+
+# Random Emojios
+EMOJIOS = [
+    "💣",
+    "💥",
+    "🪄",
+    "🧨",
+    "⚡",
+    "🤡",
+    "👻",
+    "🎃",
+    "🎩",
+    "🕊",
+]
