--- nexichat/db/__init__.py
@@ -1,2 +1,0 @@
-from .chatsdb import *
-from .usersdb import *
