--- nexichat/modules/helpers/inline.py
@@ -1,134 +1,0 @@
-from pyrogram.enums import ButtonStyle
-from pyrogram.types import InlineKeyboardButton
-
-from config import SUPPORT_GRP, UPDATE_CHNL
-from nexichat import OWNER
-from nexichat import app
-
-DEV_OP = [
-    [
-        InlineKeyboardButton(text="🥀 ᴏᴡɴᴇʀ 🥀", user_id=OWNER, style=ButtonStyle.PRIMARY),
-        InlineKeyboardButton(text="✨ ꜱᴜᴘᴘᴏʀᴛ ✨", url=f"https://t.me/{SUPPORT_GRP}", style=ButtonStyle.PRIMARY),
-    ],
-    [
-        InlineKeyboardButton(
-            text="✦ ᴀᴅᴅ ᴍᴇ ʙᴀʙʏ ✦",
-            url=f"https://t.me/{app.username}?startgroup=true",
-            style=ButtonStyle.SUCCESS
-        ),
-    ],
-    [
-        InlineKeyboardButton(text="« ʜᴇʟᴘ »", callback_data="HELP", style=ButtonStyle.PRIMARY),
-    ],
-    [
-       # InlineKeyboardButton(text="❄️ sᴏᴜʀᴄᴇ ❄️", callback_data="SOURCE", style=ButtonStyle.PRIMARY),
-        InlineKeyboardButton(text="☁️ ᴀʙᴏᴜᴛ ☁️", callback_data="ABOUT", style=ButtonStyle.DANGER),
-    ],
-]
-
-PNG_BTN = [
-    [
-        InlineKeyboardButton(
-            text="😍 ᴀᴅᴅ ᴍᴇ ʙᴀʙʏ 😍",
-            url=f"https://t.me/{app.username}?startgroup=true",
-            style=ButtonStyle.SUCCESS
-        ),
-    ],
-    [
-        InlineKeyboardButton(
-            text="⦿ ᴄʟᴏsᴇ ⦿",
-            callback_data="CLOSE",
-            style=ButtonStyle.DANGER
-        ),
-    ],
-]
-
-
-BACK = [
-    [
-        InlineKeyboardButton(text="⦿ ʙᴀᴄᴋ ⦿", callback_data="BACK", style=ButtonStyle.DEFAULT),
-    ],
-]
-
-
-HELP_BTN = [
-    [
-        InlineKeyboardButton(text="🐳 ᴄʜᴀᴛʙᴏᴛ 🐳", callback_data="CHATBOT_CMD", style=ButtonStyle.PRIMARY),
-        InlineKeyboardButton(text="🎄 ᴛᴏᴏʟs 🎄", callback_data="TOOLS_DATA", style=ButtonStyle.PRIMARY),
-    ],
-    [
-        InlineKeyboardButton(text="⦿ ʙᴀᴄᴋ ⦿", callback_data="BACK", style=ButtonStyle.DEFAULT),
-        InlineKeyboardButton(text="⦿ ᴄʟᴏsᴇ ⦿", callback_data="CLOSE", style=ButtonStyle.DANGER),
-    ],
-]
-
-
-CLOSE_BTN = [
-    [
-        InlineKeyboardButton(text="⦿ ᴄʟᴏsᴇ ⦿", callback_data="CLOSE", style=ButtonStyle.DANGER),
-    ],
-]
-
-
-CHATBOT_ON = [
-    [
-        InlineKeyboardButton(text="ᴇɴᴀʙʟᴇ", callback_data=f"addchat", style=ButtonStyle.SUCCESS),
-        InlineKeyboardButton(text="ᴅɪsᴀʙʟᴇ", callback_data=f"rmchat", style=ButtonStyle.DANGER),
-    ],
-]
-
-
-MUSIC_BACK_BTN = [
-    [
-        InlineKeyboardButton(text="sᴏᴏɴ", callback_data=f"soom", style=ButtonStyle.PRIMARY),
-    ],
-]
-
-S_BACK = [
-    [
-        InlineKeyboardButton(text="⦿ ʙᴀᴄᴋ ⦿", callback_data="SBACK", style=ButtonStyle.DEFAULT),
-        InlineKeyboardButton(text="⦿ ᴄʟᴏsᴇ ⦿", callback_data="CLOSE", style=ButtonStyle.DANGER),
-    ],
-]
-
-
-CHATBOT_BACK = [
-    [
-        InlineKeyboardButton(text="⦿ ʙᴀᴄᴋ ⦿", callback_data="CHATBOT_BACK", style=ButtonStyle.DEFAULT),
-        InlineKeyboardButton(text="⦿ ᴄʟᴏsᴇ ⦿", callback_data="CLOSE", style=ButtonStyle.DANGER),
-    ],
-]
-
-
-HELP_START = [
-    [
-        InlineKeyboardButton(text="« ʜᴇʟᴘ »", callback_data="HELP", style=ButtonStyle.PRIMARY),
-        InlineKeyboardButton(text="🐳 ᴄʟᴏsᴇ 🐳", callback_data="CLOSE", style=ButtonStyle.DANGER),
-    ],
-]
-
-
-HELP_BUTN = [
-    [
-        InlineKeyboardButton(
-            text="« ʜᴇʟᴘ »", url=f"https://t.me/{app.username}?start=help", style=ButtonStyle.PRIMARY
-        ),
-        InlineKeyboardButton(text="⦿ ᴄʟᴏsᴇ ⦿", callback_data="CLOSE", style=ButtonStyle.DANGER),
-    ],
-]
-
-
-ABOUT_BTN = [
-    [
-        InlineKeyboardButton(text="🎄 sᴜᴘᴘᴏʀᴛ 🎄", url=f"https://t.me/{SUPPORT_GRP}", style=ButtonStyle.PRIMARY),
-        InlineKeyboardButton(text="« ʜᴇʟᴘ »", callback_data="HELP", style=ButtonStyle.PRIMARY),
-    ],
-    [
-        InlineKeyboardButton(text="🍾 ᴏᴡɴᴇʀ 🍾", user_id=OWNER, style=ButtonStyle.PRIMARY),
-     #   InlineKeyboardButton(text="❄️ sᴏᴜʀᴄᴇ ❄️", callback_data="SOURCE", style=ButtonStyle.PRIMARY),
-    ],
-    [
-        InlineKeyboardButton(text="🐳 ᴜᴘᴅᴀᴛᴇs 🐳", url=f"https://t.me/{UPDATE_CHNL}", style=ButtonStyle.PRIMARY),
-        InlineKeyboardButton(text="⦿ ʙᴀᴄᴋ ⦿", callback_data="BACK", style=ButtonStyle.DEFAULT),
-    ],
-]
