+++ start
@@ -1,1 +1,1 @@
-python3 -m nexichat
+python3 -m dixitabot
